+++ packages/x-components/src/__stubs__/promoteds-stubs.factory.ts
@@ -1,66 +1,37 @@
-import { Promoted, Tagging } from '@empathyco/x-types';
-import { DeepPartial } from '../utils/types';
+import { Promoted } from '@empathyco/x-types';
 
 /**
- * Creates {@link @empathyco/x-types#Promoted | promoted} stub.
+ * Creates {@link @empathyco/x-types#Promoted | Promoted} stub.
  *
- * @returns Array of promoteds stub.
+ * @returns Array of Promoted stub.
  *
  * @internal
  */
 export function getPromotedsStub(): Promoted[] {
-  return [
-    {
-      ...getPromotedCommonValues(),
-      id: 'xp-001',
-      title: 'Promoted 01',
-      url: 'http://x-components.com',
-      image: 'xc-01.jpg',
-      tagging: {
-        click: getTaggingByAction('click', { productId: 'xc-001' })
+  return [createPromotedStub('1'), createPromotedStub('2')];
+}
+
+/**
+ * Creates a promoted with a "unique" identifier.
+ *
+ * @param identifier - The promoted identifier.
+ *
+ * @returns The promoted.
+ *
+ * @internal
+ */
+export function createPromotedStub(identifier: string): Promoted {
+  return {
+    id: `xp-${identifier}`,
+    title: `Promoted ${identifier}`,
+    url: `http://x-components-promoted-${identifier}.com`,
+    image: `xp-${identifier}.jpg`,
+    tagging: {
+      click: {
+        params: {},
+        url: ''
       }
     },
-    {
-      ...getPromotedCommonValues(),
-      id: 'xp-002',
-      title: 'Promoted 02',
-      url: 'http://x-components.com',
-      image: 'xc-02.jpg',
-      tagging: {
-        click: getTaggingByAction('click', { productId: 'xc-002' })
-      }
-    }
-  ] as Promoted[];
-}
-
-/**
- * Creates a deep partial promoted with common values for all promoted stub array.
- *
- * @returns DeepPartial promoted.
- *
- * @internal
- */
-function getPromotedCommonValues(): DeepPartial<Promoted> {
-  return {
     modelName: 'Promoted'
-  };
-}
-
-/**
- * Creates a {@link @empathyco/x-types#Tagging | promoted tagging} mocked object.
- *
- * @param action - String with the action to tag.
- * @param params - Params to add to the tagging request.
- *
- * @returns Tagging mocked object.
- *
- * @internal
- */
-function getTaggingByAction(action: string, params: Record<string, any>): Tagging {
-  return {
-    url: `http://x-components.com/tagging/${action}`,
-    params: {
-      ...params
-    }
   };
 }
