+++ packages/x-components/src/x-modules/search/components/spellcheck.vue
@@ -9,14 +9,15 @@
   import { Component } from 'vue-property-decorator';
   import { State, xComponentMixin } from '../../../components';
   import { searchXModule } from '../x-module';
+
   /**
    * The `Spellcheck` component allows to inform the user with a friendly message that he
-   might have misspelled the search query. This message can be set using the default slot
-   of the component, which gives access to the searched query using the `query` scope property,
-   and the spellchecked query proposal, using the `spellcheckedQuery` scope property.
-
-   The component will only render itself if the `spellcheckedQuery` property has value.
-
+   * might have misspelled the search query. This message can be set using the default slot
+   * of the component, which gives access to the searched query using the `query` scope property,
+   * and the spellchecked query proposal, using the `spellcheckedQuery` scope property.
+   *
+   * The component will only render itself if the `spellcheckedQuery` property has value.
+   *
    * @public
    */
   @Component({
